vcmp imp compiles
