+++ pallets/identity-management/Cargo.toml
@@ -22,7 +22,6 @@ frame-system = { git = "https://github.com/paritytech/substrate", branch = "polk
 
 # benchmarking
 frame-benchmarking = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false, optional = true }
-test-utils = { path = "../test-utils", default-features = false, optional = true }
 
 # local
 core-primitives = { path = "../../primitives/core", default-features = false }
@@ -33,7 +32,7 @@ pallet-balances = { git = "https://github.com/paritytech/substrate", branch = "p
 pallet-group = { path = "../../pallets/group" }
 pallet-timestamp = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
 sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
-test-utils = { path = "../test-utils" }
+pallet-teebag = { path = "../teebag", features = ["test-util"] }
 
 [features]
 default = ["std"]
@@ -42,7 +41,7 @@ runtime-benchmarks = [
     "frame-support/runtime-benchmarks",
     "frame-system/runtime-benchmarks",
     "pallet-timestamp/runtime-benchmarks",
-    "test-utils",
+    "pallet-teebag/runtime-benchmarks",
 ]
 skip-ias-check = [
 ]
