+++ pallets/identity-management/src/mock.rs
@@ -61,9 +61,8 @@ where
 
 	#[cfg(feature = "runtime-benchmarks")]
 	fn try_successful_origin() -> Result<T::RuntimeOrigin, ()> {
-		use test_utils::ias::consts::{TEST8_MRENCLAVE, TEST8_SIGNER_PUB};
-		let signer: <T as frame_system::Config>::AccountId =
-			test_utils::get_signer(TEST8_SIGNER_PUB);
+		use pallet_teebag::test_util::{get_signer, TEST8_MRENCLAVE, TEST8_SIGNER_PUB};
+		let signer: <T as frame_system::Config>::AccountId = get_signer(TEST8_SIGNER_PUB);
 		if !pallet_teebag::EnclaveRegistry::<T>::contains_key(signer.clone()) {
 			assert_ok!(pallet_teebag::Pallet::<T>::add_enclave(
 				&signer,
@@ -169,15 +168,18 @@ impl pallet_group::Config for Test {
 }
 
 pub fn new_test_ext() -> sp_io::TestExternalities {
+	use pallet_teebag::test_util::{
+		get_signer, TEST8_CERT, TEST8_SIGNER_PUB, TEST8_TIMESTAMP, URL,
+	};
+
 	let t = system::GenesisConfig::default().build_storage::<Test>().unwrap();
-	let eddie: SystemAccountId = test_utils::get_signer(&[5u8; 32]);
+	let eddie: SystemAccountId = get_signer(&[5u8; 32]);
 	let mut ext = sp_io::TestExternalities::new(t);
 	ext.execute_with(|| {
 		// add `5` to delegatee
 		let _ = IdentityManagement::add_delegatee(RuntimeOrigin::root(), eddie);
 		System::set_block_number(1);
-		use test_utils::ias::consts::{TEST8_CERT, TEST8_SIGNER_PUB, TEST8_TIMESTAMP, URL};
-		let signer: SystemAccountId = test_utils::get_signer(TEST8_SIGNER_PUB);
+		let signer: SystemAccountId = get_signer(TEST8_SIGNER_PUB);
 		assert_ok!(Teebag::set_admin(RuntimeOrigin::root(), signer.clone()));
 		assert_ok!(Teebag::set_mode(
 			RuntimeOrigin::signed(signer.clone()),
