+++ pallets/teebag/Cargo.toml
@@ -14,6 +14,7 @@ chrono = { version = "0.4", default-features = false, features = ["serde"] }
 codec = { version = "3.0.0", default-features = false, features = ["derive"], package = "parity-scale-codec" }
 der = { default-features = false, version = "0.6.0" }
 hex = { default-features = false, version = "0.4.3", features = ["alloc"] }
+hex-literal = { version = "0.4", default-features = false }
 log = { version = "0.4", default-features = false }
 ring = { version = "0.16.20", default-features = false, features = ["alloc"] }
 scale-info = { version = "2.10.0", default-features = false, features = ["derive"] }
@@ -31,18 +32,14 @@ sp-runtime = { default-features = false, git = "https://github.com/paritytech/su
 sp-std = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
 
 frame-benchmarking = { default-features = false, optional = true, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-hex-literal = { version = "0.4.1", optional = true }
 pallet-balances = { default-features = false, optional = true, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-test-utils = { path = "../test-utils", default-features = false, optional = true }
 
 [dev-dependencies]
 env_logger = "0.9.0"
 externalities = { package = "sp-externalities", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
 frame-benchmarking = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-hex-literal = "0.4.1"
 pallet-balances = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
 sp-keyring = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-test-utils = { path = "../test-utils" }
 
 [features]
 default = ["std"]
@@ -70,9 +67,9 @@ std = [
 ]
 runtime-benchmarks = [
     "frame-benchmarking/runtime-benchmarks",
-    "hex-literal",
     "pallet-balances?/runtime-benchmarks",
-    "test-utils",
     "pallet-timestamp/runtime-benchmarks",
 ]
 try-runtime = ["frame-support/try-runtime"]
+# workaround to cross crate boundary, see https://github.com/rust-lang/cargo/issues/8379
+test-util = []
