+++ pallets/teebag/src/sgx_verify/mod.rs
@@ -50,6 +50,7 @@ use serde_json::Value;
 use sp_std::{
 	convert::{TryFrom, TryInto},
 	prelude::*,
+	vec,
 };
 use x509_cert::Certificate;
