--- pallets/test-utils/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "test-utils"
-description = "A crate for common utilities for tests"
-version = "0.1.0"
-authors = ['Trust Computing GmbH <info@litentry.com>']
-homepage = 'https://litentry.com/'
-repository = 'https://github.com/litentry/litentry-parachain'
-license = "Apache-2.0"
-edition = "2021"
-
-[dependencies]
-hex-literal = { version = "0.4.1" }
-
-[dependencies.teerex-primitives]
-default-features = false
-package = "teerex-primitives"
-path = "../../primitives/teerex"
-
-
-[features]
-default = ['std']
-std = [
-    "teerex-primitives/std",
-]
