+++ pallets/vc-management/Cargo.toml
@@ -26,25 +26,25 @@ pallet-teebag = { path = "../teebag", default-features = false }
 
 # benchmarking
 frame-benchmarking = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false, optional = true }
-test-utils = { path = "../test-utils", default-features = false, optional = true }
 
 [dev-dependencies]
 frame-benchmarking = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
 pallet-balances = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
-pallet-group = { path = "../../pallets/group" }
 pallet-timestamp = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
 sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42" }
-test-utils = { path = "../test-utils" }
+pallet-group = { path = "../../pallets/group" }
+pallet-teebag = { path = "../teebag", features = ["test-util"] }
+
 
 [features]
 default = ["std"]
 runtime-benchmarks = [
-    "frame-benchmarking/runtime-benchmarks",
+    "frame-benchmarking?/runtime-benchmarks",
     "frame-support/runtime-benchmarks",
     "frame-system/runtime-benchmarks",
     "pallet-balances/runtime-benchmarks",
     "pallet-group/runtime-benchmarks",
-    "test-utils",
+    "pallet-teebag/runtime-benchmarks",
 ]
 std = [
     "codec/std",
