+++ pallets/vc-management/src/lib.rs
@@ -24,6 +24,7 @@
 
 #[cfg(feature = "runtime-benchmarks")]
 mod benchmarking;
+
 #[cfg(test)]
 mod mock;
