+++ runtime/rococo/Cargo.toml
@@ -93,16 +93,12 @@ pallet-bridge = { path = "../../pallets/bridge", default-features = false }
 pallet-bridge-transfer = { path = "../../pallets/bridge-transfer", default-features = false }
 pallet-drop3 = { path = "../../pallets/drop3", default-features = false }
 pallet-extrinsic-filter = { path = "../../pallets/extrinsic-filter", default-features = false }
+pallet-group = { path = "../../pallets/group", default-features = false }
 pallet-identity-management = { path = "../../pallets/identity-management", default-features = false }
 pallet-parachain-staking = { path = "../../pallets/parachain-staking", default-features = false }
+pallet-teebag = { path = "../../pallets/teebag", default-features = false }
 pallet-vc-management = { path = "../../pallets/vc-management", default-features = false }
 runtime-common = { path = '../common', default-features = false }
-
-# TEE pallets
-pallet-group = { path = "../../pallets/group", default-features = false }
-pallet-sidechain = { path = "../../pallets/sidechain", default-features = false }
-pallet-teebag = { path = "../../pallets/teebag", default-features = false }
-pallet-teerex = { path = "../../pallets/teerex", default-features = false }
 
 # EVM/Frontier
 fp-rpc = { git = "https://github.com/paritytech/frontier", branch = "polkadot-v0.9.42", default-features = false }
@@ -177,9 +173,7 @@ runtime-benchmarks = [
     "cumulus-pallet-xcmp-queue/runtime-benchmarks",
     "pallet-group/runtime-benchmarks",
     "pallet-identity-management/runtime-benchmarks",
-    "pallet-teerex/runtime-benchmarks",
     "pallet-teebag/runtime-benchmarks",
-    "pallet-sidechain/runtime-benchmarks",
     "pallet-vc-management/runtime-benchmarks",
     "pallet-account-fix/runtime-benchmarks",
 ]
@@ -267,9 +261,7 @@ std = [
     "pallet-extrinsic-filter/std",
     "pallet-group/std",
     "pallet-identity-management/std",
-    "pallet-teerex/std",
     "pallet-teebag/std",
-    "pallet-sidechain/std",
     "pallet-vc-management/std",
     "pallet-account-fix/std",
     "moonbeam-evm-tracer/std",
@@ -278,7 +270,6 @@ std = [
     "pallet-bitacross/std",
 ]
 tee-dev = [
-    "pallet-teerex/skip-ias-check",
 ]
 try-runtime = [
     "cumulus-pallet-aura-ext/try-runtime",
@@ -315,9 +306,7 @@ try-runtime = [
     "pallet-proxy/try-runtime",
     "pallet-scheduler/try-runtime",
     "pallet-session/try-runtime",
-    "pallet-sidechain/try-runtime",
     "pallet-sudo/try-runtime",
-    "pallet-teerex/try-runtime",
     "pallet-teebag/try-runtime",
     "pallet-timestamp/try-runtime",
     "pallet-tips/try-runtime",
